requirements.txt
faiss-cpu~=1.8
pandas~=2.0
numpy~=1.24
websocket-client~=1.7
open-clip-torch~=2.24
websockets~=12.0
requests~=2.31
python-dotenv~=1.0.1
opencv-python~=4.8
openai-whisper==20240930
git+https://github.com/openai/CLIP.git